requirements.txt
PyYAML >= 6.0.2
rich==13.7.1
Jinja2 >= 3.0
